ingestion/requirements.txt
warcio
beautifulsoup4
readability-lxml
requests
pandas
lxml
